Move FE components to management-data folder
RISDEV-7247
