+++ frontend/src/components/management-data/DuplicateRelationListItem.vue
