+++ frontend/src/components/management-data/ManagementDataMetadata.vue
