+++ frontend/src/routes/caselaw/documentUnit/[documentNumber]/managementdata.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import DocumentUnitManagementData from "@/components/DocumentUnitManagementData.vue"
+import DocumentUnitManagementData from "@/components/management-data/DocumentUnitManagementData.vue"
 </script>
 
 <template>
