+++ frontend/test/components/management-data/documentUnitManagementData.spec.ts
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/vue"
 import { setActivePinia } from "pinia"
 import { it } from "vitest"
 import { createRouter, createWebHistory } from "vue-router"
-import DocumentUnitManagementData from "@/components/DocumentUnitManagementData.vue"
+import DocumentUnitManagementData from "@/components/management-data/DocumentUnitManagementData.vue"
 import DocumentUnit, {
   DuplicateRelation,
   DuplicateRelationStatus,
